+++ components/catalogs/back/src/Infrastructure/Persistence/SearchAttributesQuery.php
@@ -15,6 +15,7 @@ use Akeneo\Tool\Component\StorageUtils\Repository\SearchableRepositoryInterface;
 final class SearchAttributesQuery implements SearchAttributesQueryInterface
 {
     private const ALLOWED_TYPES = [
+        'pim_catalog_identifier',
         'pim_catalog_text',
         'pim_catalog_textarea',
         'pim_catalog_simpleselect',
@@ -22,6 +23,7 @@ final class SearchAttributesQuery implements SearchAttributesQueryInterface
         'pim_catalog_number',
         'pim_catalog_metric',
         'pim_catalog_boolean',
+        'pim_catalog_date',
     ];
 
     public function __construct(