+++ components/catalogs/front/src/components/CatalogEdit/utils/mapProductValueFiltersErrors.ts
@@ -5,4 +5,5 @@ import {ProductValueFiltersErrors} from '../../ProductValueFilters';
 export const mapProductValueFiltersErrors = (errors: CatalogFormErrors): ProductValueFiltersErrors => ({
     channels: findFirstError(errors, '[product_value_filters][channels]'),
     locales: findFirstError(errors, '[product_value_filters][locales]'),
+    currencies: findFirstError(errors, '[product_value_filters][currencies]'),
 });