+++ components/catalogs/front/src/hooks/useChannelsByCodes.ts
@@ -10,7 +10,7 @@ type Result = {
 };
 
 export const useChannelsByCodes = (codes: string[]): Result => {
-    return useQuery<Channel[], ResultError, Channel[]>(['channels', codes.sort().join('')], async () => {
+    return useQuery<Channel[], ResultError, Channel[]>(['channels', [...codes].sort().join('')], async () => {
         const _codes = codes.join(',');
 
         const response = await fetch(`/rest/catalogs/channels?codes=${_codes}`, {