+++ src/Akeneo/Category/front/src/feature/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './permissionsHelper';
 export * from './treeHelper';
+export * from './categoryHelper';