+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Infrastructure/StorageClient/FileSystemStorageClient.php
@@ -48,4 +48,9 @@ final class FileSystemStorageClient implements StorageClientInterface
     {
         $this->filesystemOperator->move($sourceFilePath, $destinationFilePath);
     }
+
+    public function delete(string $filePath): void
+    {
+        $this->filesystemOperator->delete($filePath);
+    }
 }