+++ src/Akeneo/Tool/Component/Connector/Job/JobParameters/DefaultValuesProvider/SimpleXlsxExport.php
@@ -9,35 +9,32 @@ use Akeneo\Tool\Component\Batch\Job\JobParameters\DefaultValuesProviderInterface
  * DefaultParameters for simple Xlsx export
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
- * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ * @copyright 2016 Akeneo SAS (https://www.akeneo.com)
+ * @license   https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 class SimpleXlsxExport implements DefaultValuesProviderInterface
 {
-    /** @var array */
-    protected $supportedJobNames;
-
     /**
-     * @param array $supportedJobNames
+     * @param array<string> $supportedJobNames
      */
-    public function __construct(array $supportedJobNames)
-    {
-        $this->supportedJobNames = $supportedJobNames;
+    public function __construct(
+        private array $supportedJobNames,
+    ) {
     }
 
     /**
      * {@inheritdoc}
      */
-    public function getDefaultValues()
+    public function getDefaultValues(): array
     {
         return [
             'storage' => [
                 'type' => 'none',
                 'file_path' => sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'export_%job_label%_%datetime%.xlsx'
             ],
-            'withHeader'        => true,
-            'linesPerFile'      => 10000,
-            'user_to_notify'    => null,
+            'withHeader' => true,
+            'linesPerFile' => 10000,
+            'users_to_notify' => [],
             'is_user_authenticated' => false,
         ];
     }
@@ -45,7 +42,7 @@ class SimpleXlsxExport implements DefaultValuesProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function supports(JobInterface $job)
+    public function supports(JobInterface $job): bool
     {
         return in_array($job->getName(), $this->supportedJobNames);
     }